+++ server/config/webpack.config.js
@@ -25,7 +25,6 @@ module.exports = {
   },
   plugins: [
     new webpack.optimize.OccurenceOrderPlugin(),
-    new webpack.HotModuleReplacementPlugin(),
     new webpack.NoErrorsPlugin(),
     new HtmlWebpackPlugin({
       template: path.join(__dirname, '../../client/index.ejs'),